(WIP!) Calculate two-stage difference-in-differences following Gardner (2021)

(WIP!) Calculate two-stage difference-in-differences following Gardner (2021)

did3s(
  data,
  yname,
  gname,
  tname,
  time_varying,
  third_stage,
  treatment,
  cluster_var,
  time_invariant = c("1"),
  weights = NULL,
  n_bootstraps = 250,
  return_bootstrap = FALSE,
  verbose = TRUE
)

Arguments

data

The dataframe containing all the variables

yname

Variable name for outcome variable

gname

Variable name for group variable

tname

Variable name for treatment variable

time_varying

vector of variable names of time-varying covariates.

third_stage

Second stage, these should be the treatment indicator(s) (e.g. treatment variable or event-study leads/lags). Formula following fixest::feols. Use i() for factor variables, see fixest::i.

treatment

A variable that = 1 if treated, = 0 otherwise

cluster_var

What variable to cluster standard errors. This can be IDs or a higher aggregate level (state for example)

time_invariant

vector of variable names of time-invariant covariates. For factor variables, write "i.varname".

weights

Optional. Variable name for regression weights.

n_bootstraps

Optional. How many bootstraps to run. Default is 250.

return_bootstrap

Optional. Logical. Will return each bootstrap second-stage estimate to allow for manual use, e.g. percentile standard errors and empirical confidence intervals.

verbose

Optional. Logical. Should information about the two-stage procedure be printed back to the user? Default is TRUE.

bootstrap

Optional. Should standard errors be calculated using bootstrap? Default is FALSE.

Value

fixest object with adjusted standard errors (either by formula or by bootstrap). All the methods from fixest package will work, including fixest::esttable and fixest::coefplot

Examples

Load example dataset which has two treatment groups and homogeneous treatment effects

# Load Example Dataset
data("df_hom")

Static TWFE

You can run a static TWFE fixed effect model for a simple treatment indicator

static <- did2s(df_hom,
    yname = "dep_var", treatment = "treat", cluster_var = "state",
    first_stage = ~ 0 | unit + year,
    second_stage = ~ i(treat, ref=FALSE))
#> Running Two-stage Difference-in-Differences
#> • first stage formula `~ 0 | unit + year`
#> • second stage formula `~ i(treat, ref = FALSE)`
#> • The indicator variable that denotes when treatment is on is `treat`
#> • Standard errors will be clustered by `state`

fixest::esttable(static)
#>                            static
#> Dependent Var.:           dep_var
#>                                  
#> treat = TRUE    1.949*** (0.0240)
#> _______________ _________________
#> S.E. type                  Custom
#> Observations               31,000
#> R2                        0.46302
#> Adj. R2                   0.46302
#> ---
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

Event Study

Or you can use relative-treatment indicators to estimate an event study estimate

es <- did2s(df_hom,
    yname = "dep_var", treatment = "treat", cluster_var = "state",
    first_stage = ~ 0 | unit + year,
    second_stage = ~ i(rel_year, ref=c(-1, Inf)))
#> Running Two-stage Difference-in-Differences
#> • first stage formula `~ 0 | unit + year`
#> • second stage formula `~ i(rel_year, ref = c(-1, Inf))`
#> • The indicator variable that denotes when treatment is on is `treat`
#> • Standard errors will be clustered by `state`

fixest::esttable(es)
#>                                  es
#> Dependent Var.:             dep_var
#>                                    
#> rel_year = -20      0.0398 (0.0463)
#> rel_year = -19      0.0761 (0.0471)
#> rel_year = -18      0.0065 (0.0437)
#> rel_year = -17     0.0811* (0.0408)
#> rel_year = -16     -0.0039 (0.0506)
#> rel_year = -15     -0.0178 (0.0403)
#> rel_year = -14      0.0154 (0.0488)
#> rel_year = -13     -0.0211 (0.0480)
#> rel_year = -12     -0.0394 (0.0497)
#> rel_year = -11      0.0052 (0.0390)
#> rel_year = -10      0.0010 (0.0245)
#> rel_year = -9      -0.0371 (0.0288)
#> rel_year = -8      -0.0133 (0.0242)
#> rel_year = -7   -0.0808*** (0.0230)
#> rel_year = -6      -0.0076 (0.0305)
#> rel_year = -5       0.0361 (0.0242)
#> rel_year = -4       0.0199 (0.0261)
#> rel_year = -3       0.0164 (0.0285)
#> rel_year = -2      -0.0412 (0.0321)
#> rel_year = 0      1.955*** (0.0526)
#> rel_year = 1      1.923*** (0.0515)
#> rel_year = 2      1.900*** (0.0530)
#> rel_year = 3      2.002*** (0.0556)
#> rel_year = 4      2.052*** (0.0541)
#> rel_year = 5      1.942*** (0.0565)
#> rel_year = 6      1.922*** (0.0504)
#> rel_year = 7      1.959*** (0.0598)
#> rel_year = 8      1.960*** (0.0554)
#> rel_year = 9      1.919*** (0.0500)
#> rel_year = 10     1.945*** (0.0496)
#> rel_year = 11     1.945*** (0.0720)
#> rel_year = 12     1.897*** (0.0800)
#> rel_year = 13     1.940*** (0.0803)
#> rel_year = 14     1.995*** (0.0811)
#> rel_year = 15     1.989*** (0.0801)
#> rel_year = 16     1.986*** (0.0820)
#> rel_year = 17     1.911*** (0.0817)
#> rel_year = 18     1.986*** (0.0819)
#> rel_year = 19     1.896*** (0.0650)
#> rel_year = 20     1.865*** (0.0750)
#> _______________ ___________________
#> S.E. type                    Custom
#> Observations                 31,000
#> R2                          0.46359
#> Adj. R2                     0.46291
#> ---
#> Signif. codes: 0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1

# plot rel_year coefficients and standard errors
fixest::coefplot(es, keep = "rel_year::(.*)")

Example from Cheng and Hoekstra (2013)

Here's an example using data from Cheng and Hoekstra (2013)

# Castle Data
castle <- haven::read_dta("https://github.com/scunning1975/mixtape/raw/master/castle.dta")

did2s(
  data = castle,
  yname = "l_homicide",
  first_stage = ~ 0 | sid + year,
  second_stage = ~ i(post, ref=0),
  treatment = "post",
  cluster_var = "state", weights = "popwt"
)
#> Running Two-stage Difference-in-Differences
#> • first stage formula `~ 0 | sid + year`
#> • second stage formula `~ i(post, ref = 0)`
#> • The indicator variable that denotes when treatment is on is `post`
#> • Standard errors will be clustered by `state`
#> OLS estimation, Dep. Var.: l_homicide
#> Observations: 550 
#> Standard-errors: Custom 
#>         Estimate Std. Error t value Pr(>|t|)    
#> post::1 0.075142    0.03538 2.12387 0.034127 *  
#> ---
#> Signif. codes:  0 '***' 0.001 '**' 0.01 '*' 0.05 '.' 0.1 ' ' 1
#> RMSE: 263.4   Adj. R2: 0.052465