Borusyak, Jaravel, and Spiess (2021) Estimator

Treatment effect estimation and pre-trend testing in staggered adoption diff-in-diff designs with an imputation approach of Borusyak, Jaravel, and Spiess (2021)

did_imputation(
  data,
  yname,
  gname,
  tname,
  idname,
  first_stage = NULL,
  weights = NULL,
  wtr = NULL,
  horizon = NULL,
  pretrends = NULL
)

Arguments

data
A data frame

yname
String. Variable name for outcome.

gname
String. Variable name for unit-specific date of treatment (never-treated should be zero or NA)

tname
String. Variable name for calendar period

idname
String. Variable name for unique unit id

first_stage
Formula for Y(0). Formula following fixest::feols. Fixed effects specified after "|". If not specified, then just unit and time fixed effects will be used.

weights
Estimation weights for observations. This is used in estimating Y(0) and also augments treatment effect weights.

wtr
Character vector of treatment weight names (see horizon for standard static and event-study weights)

horizon
Integer vector of event_time or TRUE. This only applies if wtr is left as NULL. if specified, weighted averages/sums of treatment effects will be reported for each of these horizons separately (i.e. tau0 for the treatment period, tau1 for one period after treatment, etc.). If TRUE, all horizons are used. If wtr and horizon are null, then the static treatment effect is calculated.

pretrends
Integer vector or TRUE. Which pretrends to estimate. If TRUE, all pretrends are used.

Examples

Load example dataset which has two treatment groups and homogeneous treatment effects

# Load Example Dataset
data("df_hom")

You can run a static TWFE fixed effect model for a simple treatment indicator

did_imputation(data = df_hom, yname = "dep_var", gname = "g",
               tname = "year", idname = "unit")
#> # A tibble: 1 x 5
#>   term  estimate std.error conf.low conf.high
#>   <chr>    <dbl>     <dbl>    <dbl>     <dbl>
#> 1 treat     2.02    0.0324     1.96      2.09

Or you can use relative-treatment indicators to estimate an event study estimate

did_imputation(data = df_hom, yname = "dep_var", gname = "g",
               tname = "year", idname = "unit", horizon=TRUE)
#> # A tibble: 21 x 5
#>    term  estimate std.error conf.low conf.high
#>    <chr>    <dbl>     <dbl>    <dbl>     <dbl>
#>  1 0         2.12    0.0737     1.97      2.26
#>  2 1         1.86    0.0767     1.71      2.01
#>  3 2         1.99    0.0714     1.85      2.13
#>  4 3         2.00    0.0765     1.85      2.15
#>  5 4         1.95    0.0754     1.80      2.10
#>  6 5         2.04    0.0758     1.89      2.19
#>  7 6         2.03    0.0722     1.89      2.17
#>  8 7         2.03    0.0754     1.88      2.17
#>  9 8         1.98    0.0749     1.83      2.12
#> 10 9         2.12    0.0727     1.98      2.26
#> # … with 11 more rows

Here's an example using data from Castle (2013)

# Castle Data
castle <- haven::read_dta("https://github.com/scunning1975/mixtape/raw/master/castle.dta")

did_imputation(data = castle, yname = "l_homicide", gname = "effyear",
               first_stage = ~ 0 | sid + year,
               tname = "year", idname = "sid")
#> # A tibble: 1 x 5
#>   term  estimate std.error conf.low conf.high
#>   <chr>    <dbl>     <dbl>    <dbl>     <dbl>
#> 1 treat   0.0798    0.0609  -0.0395     0.199